e2e: move pod related query functions to libpod
this is done as part of the effort to reduce the utils.go
boilerplate.
