+++ tests/libpod/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "blackhole.go",
+        "query.go",
         "render.go",
     ],
     importpath = "kubevirt.io/kubevirt/tests/libpod",
@@ -14,6 +15,7 @@ go_library(
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
         "//tests/exec:go_default_library",
         "//tests/flags:go_default_library",
+        "//tests/framework/kubevirt:go_default_library",
         "//tests/testsuite:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
